ui changes

+++ client/src/Pages/DeletePost.js
@@ -1,13 +1,13 @@
 import { useState, useEffect } from "react";
 import { Navigate, useParams } from "react-router-dom";
-
+import Swal from "sweetalert2";
 const DeletePost = () => {
   const [redirect, setRedirect] = useState(false);
-    const {id}=useParams();
+  const { id } = useParams();
   useEffect(() => {
     const deletePost = async () => {
-        console.log('Going to delete the post!! with id ',id);
-        const response = await fetch(`http://localhost:4000/post/${id}`, {
+      console.log("Going to delete the post!! with id ", id);
+      const response = await fetch(`http://localhost:4000/post/${id}`, {
         method: "DELETE",
         credentials: "include",
       });
@@ -18,8 +18,24 @@ const DeletePost = () => {
         console.log("Post Not Deleted");
       }
     };
-
-    deletePost();
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deletePost();
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success",
+        });
+      }
+    });
   }, [id]);
 
   if (redirect) {
@@ -28,7 +44,7 @@ const DeletePost = () => {
 
   return (
     <div>
-      <h1>Post Deleted</h1>
+      <h1>Confirm</h1>
     </div>
   );
 };
